+++ projection/winrt-WindowsAppSDK/src/winrt_windows_app_sdk/cppwinrt/winrt/Microsoft.UI.Xaml.Controls.AnimatedVisuals.h
@@ -6,7 +6,7 @@
 #include "winrt/base.h"
 static_assert(winrt::check_version(CPPWINRT_VERSION, "2.0.250303.1"), "Mismatched C++/WinRT headers.");
 #define CPPWINRT_VERSION "2.0.250303.1"
-#include "winrt/Microsoft.UI.Xaml.Controls.h"
+#include "winrt/impl/Microsoft.UI.Xaml.Controls.2.h"
 #include "winrt/impl/Microsoft.UI.Xaml.Controls.2.h"
 #include "winrt/impl/Microsoft.UI.Xaml.Controls.AnimatedVisuals.2.h"
 namespace winrt::impl