+++ projection/winrt-WindowsAppSDK/src/winrt_windows_app_sdk/cppwinrt/winrt/Microsoft.Windows.ApplicationModel.Background.UniversalBGTask.h
@@ -6,7 +6,7 @@
 #include "winrt/base.h"
 static_assert(winrt::check_version(CPPWINRT_VERSION, "2.0.250303.1"), "Mismatched C++/WinRT headers.");
 #define CPPWINRT_VERSION "2.0.250303.1"
-#include "winrt/Microsoft.Windows.ApplicationModel.Background.h"
+#include "winrt/impl/Microsoft.Windows.ApplicationModel.Background.2.h"
 #include "winrt/impl/Windows.ApplicationModel.Background.2.h"
 #include "winrt/impl/Microsoft.Windows.ApplicationModel.Background.UniversalBGTask.2.h"
 namespace winrt::impl