+++ projection/winrt-sdk/src/winrt_sdk/cppwinrt/winrt/Windows.Devices.Lights.h
@@ -6,7 +6,7 @@
 #include "winrt/base.h"
 static_assert(winrt::check_version(CPPWINRT_VERSION, "2.0.250303.1"), "Mismatched C++/WinRT headers.");
 #define CPPWINRT_VERSION "2.0.250303.1"
-#include "winrt/Windows.Devices.h"
+#include "winrt/impl/Windows.Devices.2.h"
 #include "winrt/impl/Windows.Foundation.2.h"
 #include "winrt/impl/Windows.Foundation.Numerics.2.h"
 #include "winrt/impl/Windows.Storage.Streams.2.h"