+++ projection/winrt-sdk/src/winrt_sdk/cppwinrt/winrt/Windows.Devices.Pwm.Provider.h
@@ -6,7 +6,7 @@
 #include "winrt/base.h"
 static_assert(winrt::check_version(CPPWINRT_VERSION, "2.0.250303.1"), "Mismatched C++/WinRT headers.");
 #define CPPWINRT_VERSION "2.0.250303.1"
-#include "winrt/Windows.Devices.Pwm.h"
+#include "winrt/impl/Windows.Devices.Pwm.2.h"
 #include "winrt/impl/Windows.Foundation.Collections.2.h"
 #include "winrt/impl/Windows.Devices.Pwm.Provider.2.h"
 namespace winrt::impl