+++ projection/winrt-sdk/src/winrt_sdk/cppwinrt/winrt/Windows.Perception.Spatial.Preview.h
@@ -6,7 +6,7 @@
 #include "winrt/base.h"
 static_assert(winrt::check_version(CPPWINRT_VERSION, "2.0.250303.1"), "Mismatched C++/WinRT headers.");
 #define CPPWINRT_VERSION "2.0.250303.1"
-#include "winrt/Windows.Perception.Spatial.h"
+#include "winrt/impl/Windows.Perception.Spatial.2.h"
 #include "winrt/impl/Windows.Foundation.Numerics.2.h"
 #include "winrt/impl/Windows.Perception.Spatial.2.h"
 #include "winrt/impl/Windows.Perception.Spatial.Preview.2.h"