+++ projection/winrt-sdk/src/winrt_sdk/cppwinrt/winrt/Windows.Security.Authentication.Identity.Core.h
@@ -6,7 +6,7 @@
 #include "winrt/base.h"
 static_assert(winrt::check_version(CPPWINRT_VERSION, "2.0.250303.1"), "Mismatched C++/WinRT headers.");
 #define CPPWINRT_VERSION "2.0.250303.1"
-#include "winrt/Windows.Security.Authentication.Identity.h"
+#include "winrt/impl/Windows.Security.Authentication.Identity.2.h"
 #include "winrt/impl/Windows.Foundation.2.h"
 #include "winrt/impl/Windows.Foundation.Collections.2.h"
 #include "winrt/impl/Windows.Security.Authentication.Identity.Core.2.h"