+++ projection/winrt-sdk/src/winrt_sdk/cppwinrt/winrt/Windows.UI.Input.Core.h
@@ -6,7 +6,7 @@
 #include "winrt/base.h"
 static_assert(winrt::check_version(CPPWINRT_VERSION, "2.0.250303.1"), "Mismatched C++/WinRT headers.");
 #define CPPWINRT_VERSION "2.0.250303.1"
-#include "winrt/Windows.UI.Input.h"
+#include "winrt/impl/Windows.UI.Input.2.h"
 #include "winrt/impl/Windows.ApplicationModel.Core.2.h"
 #include "winrt/impl/Windows.System.2.h"
 #include "winrt/impl/Windows.UI.Core.2.h"