+++ projection/winrt-sdk/src/winrt_sdk/cppwinrt/winrt/Windows.UI.Xaml.Controls.h
@@ -6,7 +6,7 @@
 #include "winrt/base.h"
 static_assert(winrt::check_version(CPPWINRT_VERSION, "2.0.250303.1"), "Mismatched C++/WinRT headers.");
 #define CPPWINRT_VERSION "2.0.250303.1"
-#include "winrt/Windows.UI.Xaml.h"
+#include "winrt/impl/Windows.UI.Xaml.2.h"
 #include "winrt/impl/Windows.ApplicationModel.Contacts.2.h"
 #include "winrt/impl/Windows.ApplicationModel.DataTransfer.2.h"
 #include "winrt/impl/Windows.ApplicationModel.Search.2.h"